# docs/swagger.yaml
definitions:
  model.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: number
      service_name:
        maxLength: 255
        minLength: 2
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      price:
        description: целое число рублей
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: number
      service_name:
        maxLength: 255
        minLength: 2
        type: string
      start_date:
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Returns a list of all subscriptions in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new subscription record
      parameters:
      - description: Subscription Data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Deletes a subscription by its ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a subscription
      tags:
      - subscriptions
    get:
      description: Returns a subscription by its unique ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Updates an existing subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Subscription Data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a subscription
      tags:
      - subscriptions
  /subscriptions/sum:
    get:
      description: Calculates the total price of subscriptions filtered by user_id,
        service_name, start_date and end_date
      parameters:
      - description: User ID
        in: query
        name: user_id
        type: string
      - description: Service Name
        in: query
        name: service_name
        type: string
      - description: Start Date (RFC3339 format)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (RFC3339 format)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              format: float64
              type: number
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get total subscription price by filters
      tags:
      - subscriptions
  /subscriptions/user/{user_id}:
    get:
      description: Returns all subscriptions associated with a given user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscriptions by User ID
      tags:
      - subscriptions
swagger: "2.0"